// src/types/SharkPosition.ts
export interface SharkPosition {
  DateTimeUTC: string;
  shark: string;
  Transmitter: string;
  lat: number;
  lon: number;
  HPE: number;
  sex: string;
  TNL_cm: number;
  X: number;
  Y: number;
  age: number;
  daily_length: number;
  sizeclass: string;
  ageclass: string;
  daily_res: number;
}